refactor(anim): one leaf resolver, name selectors, probes out
Same behavior, fewer failure points:

* Overview::leaf(name) is now the single choke point for the
  <leaf>_enabled/_ms/_curve contract — master/group off -> 1ms,
  _ms = -1 inherits duration, unknown name falls back to duration +
  easeout instead of dereferencing a null (the old animMs() path).
  The three duplicated sentinel resolutions (anim/animMs/resolveLeaf)
  collapse into it; entryLeaf/ghostLeaf/glideLeaf shrink to pure name
  selectors.
* AnimCfg.ms is double (it always carried resolved time).
* The stagger math shared by tileProgress/tileAppear lives in one
  helper instead of two diverging copies.
* m_populate renamed m_rebuildClock — one rebuild-transition clock
  whose window entries and ghost exits read; populateMs()/dropDur()/
  animMs() helpers deleted (leaf(x).ms at the call sites).
* TEMP diagnostics from the ghost-invisibility hunt removed:
  m_dbgRWL/dbgRWLOn* + the rWL probe in renderWindowLive, Ghost/
  SwapFX::dbgLogged first-draw traces.

+++ src/anim/clocks.hpp
@@ -65,11 +65,11 @@ private:
   using clock = std::chrono::steady_clock;
 };
 
-// Snapshot of one animation leaf's config, resolved per frame by
-// Overview::anim() from cfg::anim.
+// Snapshot of one animation group's config, resolved per frame by
+// Overview::leaf() from cfg::anim.
 struct AnimCfg {
   bool on = false;
-  int ms = 1;
+  double ms = 1.0; // resolved window: >= 1, never the -1 sentinel
   std::string curve = "easeout";
 };
